requirements.txt
streamlit==1.31.0
pandas==2.1.1
openai==1.25.0
plotly==5.18.0
python-dotenv==1.0.0
numpy==1.26.0